TST: sparse.linalg: move SVDS tests to their own file
